feat(backend): Add favorite issues and some bug fixes

+++ src/controllers/answer.ts
@@ -64,7 +64,7 @@ export const sendRequest = async (req: Request, verb: string, modelName: string,
 export const getAnswerById = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
   try {
     const { id } = req.params
-    const answer = await Answer.findById(id).populate([{path: 'userId'}, {path: "issueId"}]);
+    const answer = await Answer.findById(id).populate([{path: 'userId'}, {path: "issueId"}, {path: "archives"}]);
     if (!answer) {
       res.status(404).json({ error: 'Answer not found', message: 'An Answer with the given Id doesn\'t exists'});
     } else {
@@ -79,7 +79,7 @@ export const getAnswerById = async (req: Request, res: Response, next: NextFunct
 
 export const getAllAnswers = async (req: Request, res: Response): Promise<void> => {
   try {
-    const answers: IAnswer[] = await Answer.find().populate([{path: 'userId'}, {path: "issueId"}]);
+    const answers: IAnswer[] = await Answer.find().populate([{path: 'userId'}, {path: "issueId"}, {path: "archives"}]);
     res.status(200).json({message: "Answers retrieved successfully", data: answers});
   } catch (error) {
     res.status(500).json({ error: error.message, message: 'Internal Server Error' });
@@ -91,9 +91,10 @@ export const createAnswer = async (req: Request, res: Response, next: NextFuncti
   try {
     const { userId, issueId, description, archives, upVote, downVote } = req.body;
     const validateAnswer = await answerValidation.validateAsync({ userId, issueId, description, archives, upVote, downVote })
-    const answer: IAnswer | null = await Answer.create({
+    let answer: IAnswer | null = await Answer.create({
       ...validateAnswer
     });
+    answer = await answer.populate([{path: 'userId'}, {path: "issueId"}, {path: "archives"}]);
     res.status(200).json({message: "Answer created successfully", data: answer});
   } catch (error) {
     res.status(500).json({ error: error.message, message: 'Internal Server Error' });
@@ -107,7 +108,7 @@ export const updateAnswerById = async (req: Request, res: Response, next: NextFu
     const { id } = req.params
     const { userId, issueId, description, archives, upVote, downVote } = req.body;
     const validateAnswer = await answerValidation.validateAsync({ userId, issueId, description, archives, upVote, downVote })
-    const answer: IAnswer = await Answer.findByIdAndUpdate( id,{validateAnswer},{ new: true });
+    const answer: IAnswer = await Answer.findByIdAndUpdate( id,{validateAnswer},{ new: true }).populate([{path: 'userId'}, {path: "issueId"}, {path: "archives"}]);
     res.status(200).json({message: "Answer updated successfully", data: answer});
   } catch (error) {
     res.status(500).json({ error: error.message, message: 'Internal Server Error' });
@@ -118,7 +119,7 @@ export const updateAnswerById = async (req: Request, res: Response, next: NextFu
 export const deleteAnswerById = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
   try {
     const { id } = req.params
-    const answer: IAnswer = await Answer.findByIdAndDelete(id);
+    const answer: IAnswer = await Answer.findByIdAndDelete(id).populate([{path: 'userId'}, {path: "issueId"}, {path: "archives"}]);
     res.status(200).json({message: "Answer deleted successfully", data: answer});
   } catch (error) {
     res.status(500).json({ error: error.message, message: 'Internal Server Error' });
@@ -154,7 +155,7 @@ export const downVote = async (req: Request, res: Response, next: NextFunction) 
       await sendRequest(req, "put", "vote", vote._id, { userId: userId, answerId: answerId, value: -1 })
       n_answer = await sendRequest(req, "put", "answer", answerId, { ...answer, downVote: answer.downVote + 1, upVote: answer.upVote - 1 })
     }
-    res.status(200).json({message: "Answer downVoted successfully", data: n_answer});
+    res.status(200).json({message: "Answer downVoted successfully", data: n_answer.data.data});
   } catch (error) {
     res.status(500).json({ error: error.message, message: 'Internal Server Error' });
 
@@ -189,7 +190,7 @@ export const upVote = async (req: Request, res: Response, next: NextFunction) =>
       await sendRequest(req, "put", "vote", vote._id, { userId: userId, answerId: answerId, value: 1 })
       n_answer = await sendRequest(req, "put", "answer", answerId, { ...answer, downVote: answer.downVote - 1, upVote: answer.upVote + 1 })
     }
-    res.status(200).json({message: "Answer upVoted successfully", data: n_answer});
+    res.status(200).json({message: "Answer upVoted successfully", data: n_answer.data.data});
   } catch (error) {
     res.status(500).json({ error: error.message, message: 'Internal Server Error' });
   }
